+++ examples/js/templates/tsClass.template.js
@@ -3,6 +3,8 @@ module.exports = {
     const abstract = formFields['abstract'] ? '/* abstract */ ' : '';
 
     return `${abstract}class ${formFields['className']} {
+    // ${JSON.stringify(formFields)}
+
     /**
     * @param ${formFields['comment']}
     */