+++ examples/ts/templates/tsClass.template.ts
@@ -2,6 +2,8 @@ export const generateMyAwesomeTemplate = (formFields: Record<string, string>): s
     const abstract = formFields['abstract'] ? 'abstract ' : '';
 
     return `${abstract}class ${formFields['className']} {
+    // ${JSON.stringify(formFields)}
+
     /**
     * @param ${formFields['comment']}
     */