+++ src/main/kotlin/com/github/sam0delkin/intellijpsa/activity/PsaStartupActivity.kt
@@ -1,0 +1,100 @@
+package com.github.sam0delkin.intellijpsa.activity
+
+import com.github.sam0delkin.intellijpsa.services.CompletionService
+import com.github.sam0delkin.intellijpsa.settings.Settings
+import com.github.sam0delkin.intellijpsa.statusBar.PsaStatusBarWidgetFactory
+import com.intellij.ide.util.RunOnceUtil
+import com.intellij.notification.NotificationGroupManager
+import com.intellij.notification.NotificationType
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.components.service
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.project.guessProjectDir
+import com.intellij.openapi.startup.StartupActivity
+import com.intellij.openapi.wm.impl.status.widget.StatusBarWidgetsManager
+import java.io.File
+import java.util.*
+
+class PsaStartupActivity : StartupActivity.Background {
+    private var timer: Timer? = null
+    override fun runActivity(project: Project) {
+        ApplicationManager.getApplication().invokeLater {
+            val completionService = project.service<CompletionService>()
+            val settings = completionService.getSettings()
+
+            if (!settings.pluginEnabled) {
+                val projectDir = project.guessProjectDir()
+
+                if (null === projectDir) {
+                    return@invokeLater
+                }
+
+                val scriptPath = projectDir.path + "/.psa/psa.sh"
+                val file = File(scriptPath)
+                if (file.exists() && file.canExecute()) {
+                    RunOnceUtil.runOnceForProject(project, "project_specific_autocomplete") {
+                        @Suppress("DialogTitleCapitalization")
+                        val notification = NotificationGroupManager.getInstance()
+                            .getNotificationGroup("PSA Notification")
+                            .createNotification(
+                                "Project Specific Autocomplete",
+                                "PSA configuration found in the root of your project. " +
+                                        "Would you like to enable it?",
+                                NotificationType.INFORMATION
+                            )
+                        notification.addAction(object: AnAction("Enable") {
+                            override fun actionPerformed(e: AnActionEvent) {
+                                settings.pluginEnabled = true
+                                settings.scriptPath = ".psa/psa.sh"
+
+                                scheduleUpdate(project, settings, completionService)
+                                notification.hideBalloon()
+                            }
+
+                        })
+                        notification.notify(project)
+                    }
+                }
+                return@invokeLater
+            }
+
+            val scriptDir = settings.getScriptDir()
+
+            if (null === scriptDir) {
+                return@invokeLater
+            }
+
+            ApplicationManager.getApplication().invokeLater {
+                if (null !== timer) {
+                    timer?.cancel()
+                }
+
+                scheduleUpdate(project, settings, completionService)
+            }
+        }
+    }
+
+    private fun scheduleUpdate(project: Project, settings: Settings, completionService: CompletionService) {
+        timer = Timer()
+        timer!!.schedule(object : TimerTask() {
+            override fun run() {
+                try {
+                    val info = completionService.getInfo(settings, project, settings.scriptPath!!, false)
+                    completionService.updateInfo(settings, info)
+                    completionService.lastResultSucceed = true
+                    completionService.lastResultMessage = ""
+                } catch (e: Exception) {
+                    completionService.lastResultSucceed = false
+                    completionService.lastResultMessage = e.message.toString()
+                }
+                val psaStatusBarWidgetFactory = PsaStatusBarWidgetFactory()
+                if (null === project.service<StatusBarWidgetsManager>().findWidgetFactory(PsaStatusBarWidgetFactory.WIDGET_ID)) {
+                    project.service<StatusBarWidgetsManager>().updateWidget(psaStatusBarWidgetFactory)
+                }
+                project.service<StatusBarWidgetsManager>().updateAllWidgets()
+            }
+        }, 500)
+    }
+}